% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean_coeff2.R
\name{calc_mean_coeff2}
\alias{calc_mean_coeff2}
\title{We calculate mean estimations of coefficients of the Cox modeling over repetitions for each scenario according to the three estimation approaches}
\usage{
calc_mean_coeff2(res)
}
\arguments{
\item{res}{: list from 'estimate_approach2' function}
}
\value{
the function returns a list containing:
\itemize{
\item $result: a dataset with mean coefficients for each scenario:
\item Exponential of mean of log Hazard Ratios
\item Mean of log Hazard Ratios
\item Mean of absolute bias
\item Mean relative bias
\item Percentage of success in unilateral (estimation of risk alpha if hr = 1, estimation of power if hr < 1)
\item Percentage of success in bilateral (estimation of risk alpha if hr = 1)
\item Percentage of failures (estimation of power if hr > 1)
\item $parameters: a dataset containing all the parameters for each scenario
}
}
\description{
We calculate mean estimations of coefficients of the Cox modeling over repetitions for each scenario according to the three estimation approaches
}
