% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_approach1.R
\name{estimate_approach1}
\alias{estimate_approach1}
\title{We estimate Hazard Ratios (with a Cox model) comparing each estimation approach (EMR, EMR_INSEE and EMR_INSEE_IMP) with the estimation based on the "true" data for each simulated scenario
This function should be used when all patients are in the same group ('HR' in function 'simulate_surv' should be equal to 1)}
\usage{
estimate_approach1(sim)
}
\arguments{
\item{sim}{list from 'simulate_surv' function}
}
\value{
the function returns a list containing:
\itemize{
\item $result: a dataset with:
\item columns corresponding to Hazard ratios comparing each estimation approach with true data
\item rows corresponding to each scenario
\item $parameters: a dataset containing all the parameters for each scenario
}
}
\description{
We estimate Hazard Ratios (with a Cox model) comparing each estimation approach (EMR, EMR_INSEE and EMR_INSEE_IMP) with the estimation based on the "true" data for each simulated scenario
This function should be used when all patients are in the same group ('HR' in function 'simulate_surv' should be equal to 1)
}
