% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_approach2.R
\name{estimate_approach2}
\alias{estimate_approach2}
\title{We perform Cox model analyses of the simulated data to compare the 2 groups (for example treatment) according to the estimation approach (true data, EMR, EMR_INSEE and EMR_INSEE_IMP) for each scenario and each repetition}
\usage{
estimate_approach2(sim)
}
\arguments{
\item{sim}{list from 'simulate_surv' function}
}
\value{
the function returns a list containing:
\itemize{
\item $result:  a dataset containing, for each estimation approach, each scenario and each repetition:
\item the log Hazard Ratio
\item the Hazard Ratio
\item the statistics of the Wald test
\item the p-value associated with the Wald test
\item the success in unilateral (0 = no, yes = 1) if p-value < 0.05 & HR < 1
\item the failure (0 = no, yes = 1) if p-value < 0.05 & HR > 1
\item the success in bilateral (0 = no, yes = 1) if p-value < 0.05
\item $parameters: a dataset containing all the parameters for each scenario
}
}
\description{
We perform Cox model analyses of the simulated data to compare the 2 groups (for example treatment) according to the estimation approach (true data, EMR, EMR_INSEE and EMR_INSEE_IMP) for each scenario and each repetition
}
