% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_alpha2.R
\name{plot_alpha2}
\alias{plot_alpha2}
\title{We plot alpha risk calculated over the repetitions for each scenario according to the three estimation approaches (EMR, EMR_INSEE and EMR_INSEE_IMP)
This function should be used when there is no difference in survival between the two groups ('hr' in function 'simulate_surv' should be equal to 1).}
\usage{
plot_alpha2(mean_res)
}
\arguments{
\item{mean_res}{: list from 'calc_mean_coeff2' function}
}
\value{
graphic of alpha risk
}
\description{
We plot alpha risk calculated over the repetitions for each scenario according to the three estimation approaches (EMR, EMR_INSEE and EMR_INSEE_IMP)
This function should be used when there is no difference in survival between the two groups ('hr' in function 'simulate_surv' should be equal to 1).
}
