% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta2.R
\name{plot_beta2}
\alias{plot_beta2}
\title{We plot beta risk calculated over the repetitions for each scenario according to the three estimation approaches (EMR, EMR_INSEE and EMR_INSEE_IMP)
This function should be used when there is a difference in survival between the two groups ('hr' in function 'simulate_surv' should not be equal to 1)}
\usage{
plot_beta2(mean_res)
}
\arguments{
\item{mean_res}{: containg the data set containing estimations of coefficients over repetitions for each scenarios and the data set containing scenarios parameters from 'mean_coef_ana_2gr' function}
}
\value{
graphic of beta risk
}
\description{
We plot beta risk calculated over the repetitions for each scenario according to the three estimation approaches (EMR, EMR_INSEE and EMR_INSEE_IMP)
This function should be used when there is a difference in survival between the two groups ('hr' in function 'simulate_surv' should not be equal to 1)
}
