% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bias2.R
\name{plot_bias2}
\alias{plot_bias2}
\title{We plot the mean absolute biases over the repetitions for each scenario according to the estimation approach (EMR, EMR_INSEE and EMR_INSEE_IMP)
This function should be used when there is a difference in survival between the two groups ('hr' in function 'simulate_surv' should not be equal to 1)}
\usage{
plot_bias2(mean_res)
}
\arguments{
\item{mean_res}{: list from 'calc_mean_coeff2' function}
}
\value{
graphic of mean absolute bias
}
\description{
We plot the mean absolute biases over the repetitions for each scenario according to the estimation approach (EMR, EMR_INSEE and EMR_INSEE_IMP)
This function should be used when there is a difference in survival between the two groups ('hr' in function 'simulate_surv' should not be equal to 1)
}
