% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hr1.R
\name{plot_hr1}
\alias{plot_hr1}
\title{We plot Hazard Ratios comparing each estimation approach (EMR, EMR_INSEE and EMR_INSEE_IMP) with the estimation based on the "true" data for each simulated scenario
This function should be used when all patients are in the same group ('HR' in function 'simulate_surv' should be equal to 1)}
\usage{
plot_hr1(res)
}
\arguments{
\item{res}{list from 'estimate_approach1' function}
}
\value{
graphic of Hazard Ratios
}
\description{
We plot Hazard Ratios comparing each estimation approach (EMR, EMR_INSEE and EMR_INSEE_IMP) with the estimation based on the "true" data for each simulated scenario
This function should be used when all patients are in the same group ('HR' in function 'simulate_surv' should be equal to 1)
}
