% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_surv1.R
\name{plot_surv1}
\alias{plot_surv1}
\title{We plot Kaplan-Meier survival curves according to the estimation approach (EMR, EMR_INSEE and EMR_INSEE_IMP) for each of the simulated scenarios
This function should be used when all patients are in the same group ('HR' in function 'simulate_surv' should be equal to 1)}
\usage{
plot_surv1(sim)
}
\arguments{
\item{sim}{list from 'simulate_surv' function}
}
\value{
graphic of survival curves
}
\description{
We plot Kaplan-Meier survival curves according to the estimation approach (EMR, EMR_INSEE and EMR_INSEE_IMP) for each of the simulated scenarios
This function should be used when all patients are in the same group ('HR' in function 'simulate_surv' should be equal to 1)
}
