% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siminsee-package.R
\docType{package}
\name{siminsee-package}
\alias{siminsee}
\alias{siminsee-package}
\title{siminsee: Simulate survival trial data with censoring}
\description{
Simulate survival trial data with censoring under various scenarios using hospital and INSEE administrative data.
}
\author{
\strong{Maintainer}: Clémence Léguillette \email{c-leguillette@o-lambret.fr}

Authors:
\itemize{
  \item Pierre-Yves Cren \email{p-cren@o-lambret.fr}
  \item Youenn Drouet \email{youenn.drouet@lyon.unicancer.fr}
}

}
\keyword{internal}
