% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_surv.R
\name{simulate_surv}
\alias{simulate_surv}
\title{We simulate different trials with possibly different parameters specified in arguments for each scenario.
There are 2 groups of patients, with the same number of patients in each group.}
\usage{
simulate_surv(
  scenario = 1,
  nrepeat = 1,
  nbp = 1000,
  tacc = 60,
  tfu = 60,
  lambda_d = 0.03,
  hr = 1,
  lambda_lfu1 = 0.01,
  lambda_lfu2 = 0.02,
  t_lfu2 = 36,
  hr_lfu = 1,
  pinsee0 = 0.95,
  pinsee1 = 0.95,
  seed = 2023
)
}
\arguments{
\item{scenario}{: the number of the scenarios (default 1 scenario)}

\item{nrepeat}{: a vector containing the number of repetitions for each scenario (default 1 repetition)}

\item{nbp}{: a vector containing the number of patients for each scenario (defaut 1,000 patients)}

\item{tacc}{: a vector containing the accrual duration in months for each scenario (default 60 months)}

\item{tfu}{: a vector containing the follow-up duration in months for each scenario (default 60 months)}

\item{lambda_d}{: a vector containing the rate of the exponential distribution for each scenario (default 0.03)}

\item{hr}{: a vector containing the Hazard Ratio for survival between group 0 and group 1 for each scenario (default 1)}

\item{lambda_lfu1}{: a vector containing the rate of lost to follow-up for patients until time 't_lfu2' for each scenario (default 0.01)}

\item{lambda_lfu2}{: a vector containing the rate of lost to follow-up for patients after time 't_lfu2' for each scenario (default 0.02)}

\item{t_lfu2}{: a vector containing the time at which rate of lost to follow-up changes for each scenario (default 36)}

\item{hr_lfu}{: a vector containing the Hazard ratio for lost to follow-up between group 0 and group 1 for each scenario (default 1)}

\item{pinsee0}{: a vector containing the administrative file completeness (insee) in group 0 for each scenario (default 0.95)}

\item{pinsee1}{: a vector containing the administrative file completeness (insee) in group 1 for each scenario (default 0.95)}

\item{seed}{: seed used to generate each scenario}
}
\value{
the function returns a list containing:
\itemize{
\item $dat: a dataset containing the simulated data for all the scenarios and repetitions
\item $parameters: a dataset containing all the parameters for each scenario
}
}
\description{
We simulate different trials with possibly different parameters specified in arguments for each scenario.
There are 2 groups of patients, with the same number of patients in each group.
}
