% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_outcomes.R
\name{summarize_outcomes}
\alias{summarize_outcomes}
\title{We calculate the percentages of patients with events, lost to follow-up, excluded alive or immortal according to the estimation approach (EMR, EMR_INSEE and EMR_INSEE_IMP) for each repetition of each scenario}
\usage{
summarize_outcomes(sim)
}
\arguments{
\item{sim}{list from 'simulate_surv' function}
}
\value{
the function returns a list containing:
\itemize{
\item $result: a dataset with the percentages according to the estimation approach for each repetition and each scenario:
\item percentages of patients with event
\item percentages of patients lost to follow-up
\item percentages of patients excluded alive
\item percentages of patients immortal
\item $parameters: a dataset containing all the parameters for each scenario
}
}
\description{
We calculate the percentages of patients with events, lost to follow-up, excluded alive or immortal according to the estimation approach (EMR, EMR_INSEE and EMR_INSEE_IMP) for each repetition of each scenario
}
